+++ features/sessions/impl/src/commonMain/kotlin/app/campfire/sessions/db/SessionDataSource.kt
@@ -80,4 +80,9 @@ interface SessionDataSource {
     libraryItemId: LibraryItemId,
     reported: Duration,
   )
+
+  suspend fun updatePlayMethod(
+    libraryItemId: LibraryItemId,
+    playMethod: PlayMethod,
+  )
 }